% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_tts_text.R
\name{prepare_tts_text}
\alias{prepare_tts_text}
\title{Pepare tts text}
\usage{
prepare_tts_text(text, character_ratio = 0.6)
}
\arguments{
\item{text}{The text to be coverted.}

\item{character_ratio}{The sensitivity with which lines with numbers and punctuation are excluded.}
}
\description{
This function takes
}
\examples{
cat_function()

}
\keyword{tts}
